requirements.txt
Flask==1.1.2
Flask-Caching==1.10.1
line-bot-sdk==1.19.0
